Restructure the app, prepare for the feature flags
