+++ src/components/MainView/DataView/DataView.js
@@ -2,18 +2,18 @@ import { inject } from "mobx-react";
 import { getRoot } from "mobx-state-tree";
 import { useCallback, useMemo, useState } from "react";
 import { FaQuestionCircle } from "react-icons/fa";
-import { useShortcut } from "../../sdk/hotkeys";
-import { Block, Elem } from "../../utils/bem";
-import { Icon } from "../Common/Icon/Icon";
-import { ImportButton } from "../Common/SDKButtons";
-import { Spinner } from "../Common/Spinner";
-import { Table } from "../Common/Table/Table";
-import { Tag } from "../Common/Tag/Tag";
-import { Tooltip } from "../Common/Tooltip/Tooltip";
-import * as CellViews from "./CellViews";
-import { GridView } from "./GridView/GridView";
-import { getStoredPageSize, Pagination, setStoredPageSize } from "../Common/Pagination/Pagination";
-import "./Table.styl";
+import { useShortcut } from "../../../sdk/hotkeys";
+import { Block, Elem } from "../../../utils/bem";
+import { Icon } from "../../Common/Icon/Icon";
+import { ImportButton } from "../../Common/SDKButtons";
+import { Spinner } from "../../Common/Spinner";
+import { Table } from "../../Common/Table/Table";
+import { Tag } from "../../Common/Tag/Tag";
+import { Tooltip } from "../../Common/Tooltip/Tooltip";
+import * as CellViews from "../../CellViews";
+import { GridView } from "../GridView/GridView";
+import { getStoredPageSize, Pagination, setStoredPageSize } from "../../Common/Pagination/Pagination";
+import "./DataView.styl";
 
 const injector = inject(({ store }) => {
   const { dataStore, currentView } = store;
