+++ src/stores/Tabs/tab_filter.js
@@ -1,7 +1,7 @@
 import { flow, getParent, getRoot, types } from "mobx-state-tree";
 import { toStudlyCaps } from "strman";
 import * as Filters from "../../components/Filters/types";
-import * as CellViews from "../../components/Table/CellViews";
+import * as CellViews from "../../components/CellViews";
 import { debounce } from "../../utils/debounce";
 import { isBlank, isDefined } from "../../utils/utils";
 import {
