+++ src/utils/feature-flags.js
@@ -1,6 +1,9 @@
 // Add an interactivity flag to the results to make some predictions' results be able to be automatically added to newly created annotations.
 export const FF_DEV_1621 = "ff_front_dev_1621_interactive_mode_150222_short";
 
+// Switch to page navigation
+export const FF_DEV_1470 = "ff_front_dev_1470_dm_pagination_010422_short";
+
 function getFeatureFlags() {
   return window.APP_SETTINGS?.feature_flags || {};
 }
